+++ tools/gha_osx_arm64_static.sh
@@ -25,12 +25,14 @@ unset CXXFLAGS
 unset CFLAGS
 
 # Configure.
+#
+# NOTE: turn off tutorials until the fmt+malloc compilation failure is resolved.
 CXX=clang++ CC=clang cmake -G Ninja ../ \
     -DCMAKE_PREFIX_PATH=$deps_dir \
     -DCMAKE_BUILD_TYPE=Debug \
     -DHEYOKA_BUILD_TESTS=yes \
     -DHEYOKA_WITH_MPPP=yes \
-    -DHEYOKA_BUILD_TUTORIALS=ON \
+    -DHEYOKA_BUILD_TUTORIALS=OFF \
     -DHEYOKA_WITH_SLEEF=yes \
     -DCMAKE_CXX_FLAGS_DEBUG="-g -Og" \
     -DHEYOKA_FORCE_STATIC_LLVM=yes \